Seperate base Modal from SubmitSolutionModal
